twitterBot/render_deploy/requirements.txt
tweepy
python-dotenv 